debugging ray generation
