+++ haskell-raytracer.cabal
@@ -59,7 +59,7 @@ executable haskell-raytracer
     main-is:          Main.hs
 
     -- Modules included in this executable, other than Main.
-    -- other-modules:
+    other-modules: Raytracer
 
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
@@ -67,15 +67,29 @@ executable haskell-raytracer
     -- Other library packages from which modules are imported.
     build-depends:    base ^>=4.14.1.0,
                       AC-PPM ^>=1.1.1,
-                      vector ^>=0.13.0.0,
                       linear ^>=1.21.10,
                       mtl ^>=2.2.2,
                       random ^>= 1.2.1.1,
-                      mwc-random ^>= 0.15.0.2,
-                      terminal-progress-bar ^>= 0.4.1
+                      mwc-random ^>= 0.15.0.2
 
     -- Directories containing source files.
     hs-source-dirs:   app
 
     -- Base language which the package is written in.
+    default-language: Haskell2010
+
+benchmark ray-benchmark
+    type: exitcode-stdio-1.0
+    main-is: bench/Main.hs
+    other-modules:
+        Raytracer
+    hs-source-dirs:
+        app, app/bench
+    build-depends:    base ^>=4.14.1.0,
+                      AC-PPM ^>=1.1.1,
+                      linear ^>=1.21.10,
+                      mtl ^>=2.2.2,
+                      random ^>= 1.2.1.1,
+                      mwc-random ^>= 0.15.0.2,
+                      criterion >=1.4 && <1.6
     default-language: Haskell2010
